+++ lib/ktropf_solver.m
@@ -93,13 +93,13 @@ bfeq = u(ieq);
 %% build admittance matrices
 [Ybus, Yf, Yt] = makeYbus(baseMVA, bus, branch);
 
-%% try to select an interior initial point
-if mpopt.opf.init_from_mpc ~= 1
+%% try to select an interior initial point, unless requested not to
+if mpopt.opf.start < 2
     s = 1e-3;                   %% set init point inside bounds by s
-    ll = xmin; uu = xmax;
-    ll(xmin == -Inf) = -1e10;   %% replace Inf with numerical proxies
-    uu(xmax ==  Inf) =  1e10;
-    x0 = (ll + uu) / 2;         %% set x0 mid-way between bounds
+    lb = xmin; ub = xmax;
+    lb(xmin == -Inf) = -1e10;   %% replace Inf with numerical proxies
+    ub(xmax ==  Inf) =  1e10;
+    x0 = (lb + ub) / 2;         %% set x0 mid-way between bounds
     k = find(xmin == -Inf & xmax < Inf);    %% if only bounded above
     x0(k) = xmax(k) - s;                    %% set just below upper bound
     k = find(xmin > -Inf & xmax == Inf);    %% if only bounded below
@@ -108,11 +108,8 @@ if mpopt.opf.init_from_mpc ~= 1
     x0(vv.i1.Va:vv.iN.Va) = Varefs(1);  %% angles set to first reference angle
     if ny > 0
         ipwl = find(gencost(:, MODEL) == PW_LINEAR);
-    %     PQ = [gen(:, PMAX); gen(:, QMAX)];
-    %     c = totcost(gencost(ipwl, :), PQ(ipwl));
         c = gencost(sub2ind(size(gencost), ipwl, NCOST+2*gencost(ipwl, NCOST)));    %% largest y-value in CCV data
         x0(vv.i1.y:vv.iN.y) = max(c) + 0.1 * abs(max(c));
-    %     x0(vv.i1.y:vv.iN.y) = c + 0.1 * abs(c);
     end
 end