+++ lib/t/t_opf_ipopt.m
@@ -13,7 +13,7 @@ if nargin < 1
     quiet = 0;
 end
 
-num_tests = 168;
+num_tests = 204;
 
 t_begin(num_tests, quiet);
 
@@ -65,20 +65,24 @@ if have_fcn('ipopt')
     load soln9_opf;     %% defines bus_soln, gen_soln, branch_soln, f_soln
 
     %% run OPF
-    t = t0;
-    [baseMVA, bus, gen, gencost, branch, f, success, et] = runopf(casefile, mpopt);
-    t_ok(success, [t 'success']);
-    t_is(f, f_soln, 3, [t 'f']);
-    t_is(   bus(:,ib_data   ),    bus_soln(:,ib_data   ), 10, [t 'bus data']);
-    t_is(   bus(:,ib_voltage),    bus_soln(:,ib_voltage),  3, [t 'bus voltage']);
-    t_is(   bus(:,ib_lam    ),    bus_soln(:,ib_lam    ),  3, [t 'bus lambda']);
-    t_is(   bus(:,ib_mu     ),    bus_soln(:,ib_mu     ),  2, [t 'bus mu']);
-    t_is(   gen(:,ig_data   ),    gen_soln(:,ig_data   ), 10, [t 'gen data']);
-    t_is(   gen(:,ig_disp   ),    gen_soln(:,ig_disp   ),  3, [t 'gen dispatch']);
-    t_is(   gen(:,ig_mu     ),    gen_soln(:,ig_mu     ),  3, [t 'gen mu']);
-    t_is(branch(:,ibr_data  ), branch_soln(:,ibr_data  ), 10, [t 'branch data']);
-    t_is(branch(:,ibr_flow  ), branch_soln(:,ibr_flow  ),  3, [t 'branch flow']);
-    t_is(branch(:,ibr_mu    ), branch_soln(:,ibr_mu    ),  2, [t 'branch mu']);
+    for s = 0:3
+        mpopt = mpoption(mpopt, 'opf.start', s);
+        t = sprintf('%s(start=%d): ', t0, s);
+        [baseMVA, bus, gen, gencost, branch, f, success, et] = runopf(casefile, mpopt);
+        t_ok(success, [t 'success']);
+        t_is(f, f_soln, 3, [t 'f']);
+        t_is(   bus(:,ib_data   ),    bus_soln(:,ib_data   ), 10, [t 'bus data']);
+        t_is(   bus(:,ib_voltage),    bus_soln(:,ib_voltage),  3, [t 'bus voltage']);
+        t_is(   bus(:,ib_lam    ),    bus_soln(:,ib_lam    ),  3, [t 'bus lambda']);
+        t_is(   bus(:,ib_mu     ),    bus_soln(:,ib_mu     ),  2, [t 'bus mu']);
+        t_is(   gen(:,ig_data   ),    gen_soln(:,ig_data   ), 10, [t 'gen data']);
+        t_is(   gen(:,ig_disp   ),    gen_soln(:,ig_disp   ),  3, [t 'gen dispatch']);
+        t_is(   gen(:,ig_mu     ),    gen_soln(:,ig_mu     ),  3, [t 'gen mu']);
+        t_is(branch(:,ibr_data  ), branch_soln(:,ibr_data  ), 10, [t 'branch data']);
+        t_is(branch(:,ibr_flow  ), branch_soln(:,ibr_flow  ),  3, [t 'branch flow']);
+        t_is(branch(:,ibr_mu    ), branch_soln(:,ibr_mu    ),  2, [t 'branch mu']);
+    end
+    mpopt = mpoption(mpopt, 'opf.start', 0);    %% set 'opf.start' back to default
 
     %% run with automatic conversion of single-block pwl to linear costs
     t = [t0 '(single-block PWL) : '];