+++ CHANGES.md
@@ -12,6 +12,11 @@ For change history for [MOST][3], see [most/CHANGES.md](most/CHANGES.md).
 Since last release
 ------------------
 
+#### 11/22/17
+  - Add new option `opf.start` to replace deprecated `opf.init_from_mpc`
+    and add a new possibility to automatically run a power flow to
+    initialize the starting state for the OPF.
+
 #### 11/21/17
   - Add MATLAB toolbox (.mltbx) version of MATPOWER to MATLAB Central
     File Exchange and update installation instructions to include option